Made the root element parent to all subsequent elements.
The form now uses //mods:mods[1] as the root element, this allows MODS documents
that use the mods:modsCollection element as the root element to work with the
form. All references to the document context save the root element have been
changed to a parent context.

This is required to use the Video Form with the MARCXML generated MODS.